refactor : 리뷰 피드백 반영
- LineTest 클래스 이름 LineAcceptanceTest로 변경
- 테스트 환경 관련 코드 AcceptanceTest 상위 클래스로 이관
